Huge CSS changes. Drop Semantic-UI.
Semantic-UI is a huge CSS framework and I only used a few lines of it.
Replicate what I did use in app.scss and also pull in Font-Awesome.

Also rewrite the site header to use flexbox and be mobile friendly, woo!

Still need to investigate Autoprefixer to make this work in all browsers.